requirements.txt
langgraph
langchain-openai
arxiv
tenacity
pydantic
python-dotenv
openai
graphviz # Для визуализации графа (опционально)